Start to refactor all of tape
Use thor, and be more OO
