+++ .gitignore
@@ -4,3 +4,8 @@ site_vars.yml
 hosts
 *.gem
 .tape
+
+
+# TODO: Remove this after refactor
+old_lib
+bin/old_tape
