+++ taperole.gemspec
@@ -1,6 +1,6 @@
 Gem::Specification.new do |spec|
   spec.name = "taperole"
-  spec.version = '1.6.0'
+  spec.version = '1.7.0'
   spec.authors = ['Jack Forrest', 'Smashing Boxes', 'Brandon Mathis']
   spec.description = "General purpose server provisioning and application deployment toolkit"
   spec.email = ['jack@smashingboxes.com', 'brandon@sbox.es']
